+++ .iex.exs
@@ -5,18 +5,18 @@
 |> IO.inspect()
 
 defmodule Receiver do
-  use Tello.Client.Receiver
+  use Tello.Controller.Receiver
 
   def receive_message(data) do
     IO.inspect("receives message from Tello: #{data}")
   end
 end
 
-{:ok, supervisor, client, status_listener} =
+{:ok, supervisor, controller, status_listener} =
    Tello.start(
-      client: [ip: {127, 0, 0, 1}, port: tello_server_port, receiver: Receiver]
+      controller: [ip: {127, 0, 0, 1}, port: tello_server_port, receiver: Receiver]
     )
 |> IO.inspect()
 
-Tello.Command.enable(client)
-Tello.Command.takeoff(client)
+Tello.Controller.enable(controller)
+Tello.Controller.takeoff(controller)