+++ lib/tello/controller/receiver.ex
@@ -1,10 +1,10 @@
-defmodule Tello.Client.Receiver do
+defmodule Tello.Controller.Receiver do
   @type t :: module()
   @callback receive_message(data :: binary()) :: none()
 
   defmacro __using__(_) do
     quote do
-      @behaviour Tello.Client.Receiver
+      @behaviour Tello.Controller.Receiver
     end
   end
 end