+++ lib/tello/controller/state.ex
@@ -1,0 +1,4 @@
+defmodule Tello.Controller.State do
+  @moduledoc false
+  defstruct [:socket, :tello_server, :receiver_module]
+end