+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## main
 
 
-- Added `Tello.Client.StatusListener` to receive status update from Tello.
-- Allow to attach custom receiver to `Tello.Client` to receive message form Tello, see `Tello.Client.Receiver` for more detail.
+- Added `Tello.StatusListener` to receive status update from Tello.
+- Allow to attach custom receiver to `Tello.Controller` to receive message form Tello, see `Tello.Controller.Receiver` for more detail.
 
 ## V0.1.0